Add unit test to determine that healtcheck outcome is up move some unit tests to more logical positions some cleanup
