+++ record-jar-converter-web/pom.xml
@@ -99,6 +99,16 @@
             <groupId>org.wildfly.swarm</groupId>
             <artifactId>microprofile-health</artifactId>
         </dependency>
+
+        <!-- Test dependencies -->
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-engine</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.rest-assured</groupId>
+            <artifactId>rest-assured</artifactId>
+        </dependency>
     </dependencies>
 
     <profiles>
@@ -112,11 +122,19 @@
                         <configuration>
                             <images>
                                 <image>
-                                    <name>postremus/${project.artifactId}</name>
+                                    <name>${docker.repository}/${project.artifactId}</name>
                                     <run>
                                         <ports>
                                             <port>8080:8080</port>
                                         </ports>
+                                        <wait>
+                                            <http>
+                                                <url>http://127.0.0.1:8080/health</url>
+                                                <method>GET</method>
+                                                <status>200...599</status>
+                                            </http>
+                                            <time>50000</time>
+                                        </wait>
                                     </run>
                                     <build>
                                         <dockerFileDir>${project.basedir}/src/main/docker</dockerFileDir>
